Add istio special tweaks due to buggy helm chart
Catch 22 of creating istio-operator namespace inside the helm chart
Workaround by annotating/labling it properly in the helmchart

Minor tweak of not blocking flux installation on deployments.
This so all manifests can be installed and in OCP we can post fix the
issue with flux running as a user presented as a string and not uid.
After first run flux will adjust this with the kustomization-override

Set kubeconfig to a static value for helm to be able to install.